role-assignment entities pkg restructure
