+++ src/integrationTest/java/uk/gov/hmcts/reform/wataskconfigurationapi/auth/role/entities/roleassignment/QueryRequestTest.java
@@ -7,10 +7,9 @@ import org.springframework.boot.test.autoconfigure.json.JsonTest;
 import org.springframework.boot.test.json.JacksonTester;
 import org.springframework.boot.test.json.JsonContent;
 import org.springframework.test.context.junit4.SpringRunner;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.QueryRequest;
 import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.RoleAttributeDefinition;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.RoleName;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.RoleType;
+import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.enums.RoleType;
+import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.request.QueryRequest;
 
 import java.time.LocalDateTime;
 import java.util.Collections;
@@ -29,7 +28,7 @@ class QueryRequestTest {
     public void testSerializeQueryRequest() throws Exception {
         QueryRequest queryRequest = QueryRequest.builder()
             .roleType(Collections.singletonList(RoleType.CASE))
-            .roleName(Collections.singletonList(RoleName.TRIBUNAL_CASEWORKER))
+            .roleName(Collections.singletonList("tribunal-caseworker"))
             .validAt(LocalDateTime.parse("2020-10-06T17:00:00"))
             .attributes(Map.of(RoleAttributeDefinition.CASE_ID.value(), Collections.singletonList("1604584759556245")))
             .build();
