+++ src/main/java/uk/gov/hmcts/reform/wataskconfigurationapi/auth/role/RoleAssignmentService.java
@@ -6,11 +6,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
 import uk.gov.hmcts.reform.wataskconfigurationapi.auth.idam.IdamTokenGenerator;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.QueryRequest;
 import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.RoleAssignment;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.RoleAssignmentResource;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.RoleName;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.RoleType;
+import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.enums.RoleType;
+import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.request.QueryRequest;
+import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.response.RoleAssignmentResource;
 import uk.gov.hmcts.reform.wataskconfigurationapi.clients.RoleAssignmentServiceApi;
 import uk.gov.hmcts.reform.wataskconfigurationapi.exceptions.ServerErrorException;
 
@@ -66,7 +65,7 @@ public class RoleAssignmentService {
     private QueryRequest buildQueryRequest(String caseId) {
         return QueryRequest.builder()
             .roleType(singletonList(RoleType.CASE))
-            .roleName(singletonList(RoleName.TRIBUNAL_CASEWORKER))
+            .roleName(singletonList("tribunal-caseworker"))
             .validAt(LocalDateTime.now())
             .attributes(singletonMap("caseId", singletonList(caseId)))
             .build();
