+++ src/main/java/uk/gov/hmcts/reform/wataskconfigurationapi/auth/role/entities/enums/Classification.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities;
+package uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.enums;
 
 import com.fasterxml.jackson.annotation.JsonEnumDefaultValue;
