+++ src/test/java/uk/gov/hmcts/reform/wataskconfigurationapi/auth/role/entities/enums/ActorIdTypeTest.java
@@ -1,8 +1,8 @@
-package uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.roleassignment;
+package uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.enums;
 
 import org.junit.jupiter.api.Test;
 import pl.pojo.tester.api.assertion.Method;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.ActorIdType;
+import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.enums.ActorIdType;
 
 import static pl.pojo.tester.api.assertion.Assertions.assertPojoMethodsFor;
