+++ src/test/java/uk/gov/hmcts/reform/wataskconfigurationapi/auth/role/entities/enums/ClassificationTest.java
@@ -1,21 +1,20 @@
-package uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.roleassignment;
+package uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.enums;
 
 import org.junit.jupiter.api.Test;
 import pl.pojo.tester.api.assertion.Method;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.RoleCategory;
+import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.enums.Classification;
 
 import static pl.pojo.tester.api.assertion.Assertions.assertPojoMethodsFor;
 
-class RoleCategoryTest {
+class ClassificationTest {
 
     @Test
     void isWellImplemented() {
-        final Class<?> classUnderTest = RoleCategory.class;
+        final Class<?> classUnderTest = Classification.class;
 
         assertPojoMethodsFor(classUnderTest)
             .testing(Method.GETTER)
             .areWellImplemented();
     }
-
 
 }
