+++ src/test/java/uk/gov/hmcts/reform/wataskconfigurationapi/auth/role/entities/enums/GrantTypeTest.java
@@ -1,20 +1,20 @@
-package uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.roleassignment;
+package uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.enums;
 
 import org.junit.jupiter.api.Test;
 import pl.pojo.tester.api.assertion.Method;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.RoleName;
+import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.enums.GrantType;
 
 import static pl.pojo.tester.api.assertion.Assertions.assertPojoMethodsFor;
 
-class RoleNameTest {
+class GrantTypeTest {
+
     @Test
     void isWellImplemented() {
-        final Class<?> classUnderTest = RoleName.class;
+        final Class<?> classUnderTest = GrantType.class;
 
         assertPojoMethodsFor(classUnderTest)
             .testing(Method.GETTER)
             .areWellImplemented();
     }
-
 
 }
