+++ src/test/java/uk/gov/hmcts/reform/wataskconfigurationapi/auth/role/entities/enums/RoleCategoryTest.java
@@ -1,20 +1,21 @@
-package uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.roleassignment;
+package uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.enums;
 
 import org.junit.jupiter.api.Test;
 import pl.pojo.tester.api.assertion.Method;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.GrantType;
+import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.enums.RoleCategory;
 
 import static pl.pojo.tester.api.assertion.Assertions.assertPojoMethodsFor;
 
-class GrantTypeTest {
+class RoleCategoryTest {
 
     @Test
     void isWellImplemented() {
-        final Class<?> classUnderTest = GrantType.class;
+        final Class<?> classUnderTest = RoleCategory.class;
 
         assertPojoMethodsFor(classUnderTest)
             .testing(Method.GETTER)
             .areWellImplemented();
     }
+
 
 }
