+++ src/test/java/uk/gov/hmcts/reform/wataskconfigurationapi/auth/role/entities/enums/RoleTypeTest.java
@@ -1,20 +1,21 @@
-package uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.roleassignment;
+package uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.enums;
 
 import org.junit.jupiter.api.Test;
 import pl.pojo.tester.api.assertion.Method;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.Classification;
+import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.enums.RoleType;
 
 import static pl.pojo.tester.api.assertion.Assertions.assertPojoMethodsFor;
 
-class ClassificationTest {
+class RoleTypeTest {
 
     @Test
     void isWellImplemented() {
-        final Class<?> classUnderTest = Classification.class;
+        final Class<?> classUnderTest = RoleType.class;
 
         assertPojoMethodsFor(classUnderTest)
             .testing(Method.GETTER)
             .areWellImplemented();
     }
+
 
 }
