+++ src/test/java/uk/gov/hmcts/reform/wataskconfigurationapi/auth/role/entities/response/RoleAssignmentResourceTest.java
@@ -1,21 +1,20 @@
-package uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.roleassignment;
+package uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.response;
 
 import org.junit.jupiter.api.Test;
 import pl.pojo.tester.api.assertion.Method;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.RoleAssignment;
+import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.response.RoleAssignmentResource;
 
 import static pl.pojo.tester.api.assertion.Assertions.assertPojoMethodsFor;
 
-class RoleAssignmentTest {
+class RoleAssignmentResourceTest {
 
     @Test
     void isWellImplemented() {
-        final Class<?> classUnderTest = RoleAssignment.class;
+        final Class<?> classUnderTest = RoleAssignmentResource.class;
 
         assertPojoMethodsFor(classUnderTest)
             .testing(Method.GETTER)
             .testing(Method.CONSTRUCTOR)
-            .testing(Method.TO_STRING)
             .testing(Method.EQUALS)
             .testing(Method.HASH_CODE)
             .areWellImplemented();
