+++ src/test/java/uk/gov/hmcts/reform/wataskconfigurationapi/services/TaskAutoAssignmentServiceTest.java
@@ -6,12 +6,11 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.RoleAssignmentService;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.ActorIdType;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.Classification;
 import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.RoleAssignment;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.RoleCategory;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.RoleName;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.RoleType;
+import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.enums.ActorIdType;
+import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.enums.Classification;
+import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.enums.RoleCategory;
+import uk.gov.hmcts.reform.wataskconfigurationapi.auth.role.entities.enums.RoleType;
 import uk.gov.hmcts.reform.wataskconfigurationapi.domain.entities.configuration.AutoAssignmentResult;
 import uk.gov.hmcts.reform.wataskconfigurationapi.domain.entities.configuration.TaskToConfigure;
 
@@ -72,7 +71,7 @@ class TaskAutoAssignmentServiceTest {
             .id("someId")
             .actorIdType(ActorIdType.IDAM)
             .actorId("someUserId")
-            .roleName(RoleName.TRIBUNAL_CASEWORKER)
+            .roleName("tribunal-caseworker")
             .roleCategory(RoleCategory.LEGAL_OPERATIONS)
             .roleType(RoleType.ORGANISATION)
             .classification(Classification.PUBLIC)
@@ -111,7 +110,7 @@ class TaskAutoAssignmentServiceTest {
             .id("someId")
             .actorIdType(ActorIdType.IDAM)
             .actorId("someUserId")
-            .roleName(RoleName.TRIBUNAL_CASEWORKER)
+            .roleName("tribunal-caseworker")
             .roleCategory(RoleCategory.LEGAL_OPERATIONS)
             .roleType(RoleType.ORGANISATION)
             .classification(Classification.PUBLIC)
